Sql storage

+++ src/pages/create-custom-form/create-custom-form.ts
@@ -1,4 +1,4 @@
-import { Component, NgZone } from '@angular/core';
+import { Component } from '@angular/core';
 import { IonicPage, NavController, NavParams, Platform, ModalController, ViewController, ActionSheetController} from 'ionic-angular';
 import { getRepository, Repository } from 'typeorm';
 import { Toast } from '@ionic-native/toast';
@@ -8,13 +8,14 @@ import { QuestionControlService }  from '../../providers/questions/question-cont
 import { FormsProvider }  from '../../providers/forms/forms';
 import { ConfigProvider }  from '../../providers/config/config';
 
+import { UtilsProvider } from '../../providers/utils/utils';
 
 import { AlertController } from 'ionic-angular';
 
 import {CustomForm} from "../../entities/customForm";
 import {CustomFormElement} from "../../entities/customFormElement";
+import {Survey} from "../../entities/survey";
 
-import { FormControl } from '@angular/forms';
 import { FormGroup } from '@angular/forms';
 
 import {ModalCreateFormFieldPage} from '../modal-create-form-field/modal-create-form-field';
@@ -31,6 +32,8 @@ export class CreateCustomFormPage {
   formElementsRepository:any;
   formEntity:CustomForm;
 
+  surveyRepository:any;
+
   form:FormGroup;
   formsList:CustomForm[] = [];
 
@@ -38,8 +41,14 @@ export class CreateCustomFormPage {
 
   contextData = {};
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, service: QuestionService, private qcs: QuestionControlService, public alertCtrl: AlertController, private modalController: ModalController, public viewCtrl: ViewController,  private toast: Toast, public platform: Platform, public actionsheetCtrl: ActionSheetController, private zone: NgZone, private formsProvider: FormsProvider, private configProvider: ConfigProvider) {
+  constructor(public navCtrl: NavController, public navParams: NavParams, service: QuestionService, 
+              public alertCtrl: AlertController, private modalController: ModalController, public viewCtrl: ViewController,  
+              private toast: Toast, public platform: Platform, public actionsheetCtrl: ActionSheetController, 
+              private formsProvider: FormsProvider, private configProvider: ConfigProvider,
+              private utils: UtilsProvider
+    ) {
     this.formRepository = getRepository('customForm') as Repository<CustomForm>;
+    this.surveyRepository = getRepository('survey') as Repository<Survey>;
     this.formElementsRepository = getRepository('customFormElement') as Repository<CustomFormElement>;
     this.formEntity = this.navParams.get("form");
     if (!this.isEditingContext()){
@@ -197,17 +206,23 @@ export class CreateCustomFormPage {
         {
           text: 'Eliminar',
           handler: () => {
-              this.formElementsRepository.remove(this.formEntity.form_elements);
-              this.formRepository.remove(this.formEntity).then(entity => {
-              self.toast.showShortTop("Formulario eliminado con éxito").subscribe(
-                entity => {
-                   if (!toastFiredOnce){
-                     self.navCtrl.popToRoot();
-                     toastFiredOnce = true;
-                   }
-                }   
-              );
-            });
+              if (this.formHasChilds()){
+                  this.utils.showBasicAlertMessage("Error", "No se puede eliminar el formulario ya que tiene otros formularios asociados (formularios hijos)");
+              } else if ( this.formHasSurveys()){
+                this.utils.showBasicAlertMessage("Error", "No se puede eliminar el formulario ya que tiene relevamientos asociados!");
+              } else {
+                  this.formElementsRepository.remove(this.formEntity.form_elements);
+                  this.formRepository.remove(this.formEntity).then(entity => {
+                  self.toast.showShortTop("Formulario eliminado con éxito").subscribe(
+                    entity => {
+                      if (!toastFiredOnce){
+                        self.navCtrl.popToRoot();
+                        toastFiredOnce = true;
+                      }
+                    }   
+                  );
+                });
+              }
           }
         }
       ]
@@ -216,6 +231,17 @@ export class CreateCustomFormPage {
     }
   }
 
+
+  formHasChilds(){
+    return this.formEntity.child_forms.length;
+  }
+
+  formHasSurveys(){
+    this.surveyRepository.find({formId: this.formEntity.id}).then( (surveys) => {
+      return surveys.length;
+    })
+
+  }
 
 
 }
